Alterado logica para usar estado global usando context.
