+++ project-pokedex/src/App.js
@@ -1,13 +1,14 @@
 import './App.css';
 import React from 'react';
 import Router from './routes/Router';
+import GlobalState from './global/GlobalState';
 
 const App = () => {
   
   return (
-    <div>
+    <GlobalState>
       <Router />
-    </div>
+    </GlobalState>
   );
 }
